refactor(cli): code
