+++ lib/utils/findPort.js
@@ -1,15 +1,15 @@
 'use strict';
 
 const { getPortPromise } = require('portfinder');
+const defaultPort = require('./defaultPort');
 const defaultTo = require('./defaultTo');
 const tryParseInt = require('./tryParseInt');
-
-const defaultPort = 8080;
 
 function findPort(port) {
   if (typeof port !== 'undefined') {
     return Promise.resolve(port);
   }
+
   // Try to find unused port and listen on it for 3 times,
   // if port is not specified in options.
   // Because NaN == null is false, defaultTo fails if parseInt returns NaN
